pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "urnc*",
]

[project]
name = "urnc"
version = "2.3.0"
authors = [
    { name = "Michael Huttner", email = "michael@mhuttner.com" },
    { name = "Tobias Schmidt", email = "tobias.schmidt331@gmail.com" },
]
description = "A package to convert UR FIDS jupyter notebooks"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click >= 8.1.3",
    "python-dateutil >= 2.8.0",
    "GitPython >= 3.1.31",
    "nbconvert >= 7.16.6",
    "semver >= 3.0.1",
    "ruamel.yaml >= 0.18.4",
    "requests >= 2.31.0",
    "papermill >= 2.4.0",
    "ipykernel >= 6.29.5"
]
[project.optional-dependencies]
dev = [
    "sphinx >= 4.0.0",
    "sphinx_rtd_theme >= 3.0.2",
    "myst_parser >= 0.18.0",
    "toml >= 0.10.2",
    "pytest >= 8.3.5",
    "pytest-cov >= 6.1.1",
    "freezegun >= 1.5.1",
    "PyYAML >= 6.0.2",
    "pytest-xdist >= 3.6.1",
    "autopep8 >= 2.3.2"
]

[project.scripts]
urnc = "urnc.main:main"

[project.urls]
Homepage = "https://spang-lab.github.io/urnc"
Source = "https://github.com/spang-lab/urnc"
"Bug Tracker" = "https://github.com/spang-lab/urnc/issues"

[tool.pytest.ini_options]
filterwarnings = [
  'ignore:.*Jupyter is migrating its paths',
  'ignore:.*utcnow\(\) is deprecated'
]
